% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{indice_n}
\alias{indice_n}
\title{indice_n}
\usage{
indice_n(x, n, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{rabarbRa object}

\item{n}{n samples}

\item{replace}{boolean}

\item{prob}{prob transition matrix}
}
\description{
indice_n
}
\examples{
\dontrun{
indice_n()
}
}
\keyword{rabarbRa}

