% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{queries}
\alias{queries}
\title{queries}
\usage{
queries(..., logic = NULL)
}
\arguments{
\item{...}{queries}

\item{logic}{logical character "AND" "OR"}
}
\description{
queries
}
\examples{
\dontrun{
multiquery()
}
}
\keyword{rabarbRa}

