% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{query}
\alias{query}
\title{query}
\usage{
query(field, fun_sign, value)
}
\arguments{
\item{field}{field}

\item{fun_sign}{fun_sign}

\item{value}{value}
}
\description{
query
}
\examples{
\dontrun{
query()
}
}
\keyword{rabarbRa}

