% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rabarbRa.R
\name{rabarbRa}
\alias{rabarbRa}
\title{rabarbRa}
\usage{
rabarbRa(df = NULL, filename = NULL, url = NULL)
}
\arguments{
\item{df}{data.frame}

\item{filename}{filename}

\item{url}{url}
}
\description{
rabarbRa
}
\examples{
\dontrun{
rabarbRa()
}
}
\keyword{rabarbRa}

