% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{as.lazy}
\alias{f_eval}
\alias{reexports}
\alias{uq}
\title{Objects exported from other packages}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{lazyeval}{\code{\link[lazyeval]{f_eval}}, \code{\link[lazyeval]{uq}}, \code{\link[lazyeval]{as.lazy}}}
}}
\keyword{internal}

