% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample.R
\name{sample_n}
\alias{sample_n}
\title{sample_row}
\usage{
sample_n(x, n, replace = FALSE, prob = NULL)
}
\arguments{
\item{x}{rabarbRa object}

\item{n}{n samples}

\item{replace}{boolean}

\item{prob}{prob transition matrix}
}
\description{
sample_row
}
\examples{
\dontrun{
select_row()
}
}
\keyword{rabarbRa}

