% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{select}
\usage{
select(x, subset, variable)
}
\arguments{
\item{x}{rabarbRa object}

\item{subset}{logical expression indicating elements or rows to keep: missing values are taken as false}

\item{variable}{expression, indicating variables to select.}
}
\description{
select
}
\examples{
\dontrun{
select()
}
}
\keyword{rabarbRa}

