% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_row}
\alias{select_row}
\title{select_row}
\usage{
select_row(x, query)
}
\arguments{
\item{x}{rabarbRa object}

\item{query}{parameters regarding the selection method}
}
\description{
select_row
}
\examples{
\dontrun{
select_row()
}
}
\keyword{rabarbRa}

