% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select_var}
\alias{select_var}
\title{select_var}
\usage{
select_var(x, variable)
}
\arguments{
\item{x}{rabarbRa object}

\item{variable}{parameters regarding the selection method}
}
\description{
select_var
}
\examples{
\dontrun{
select_var()
}
}
\keyword{rabarbRa}

